requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
Pillow>=10.2.0
aiosqlite>=0.19.0
aiofiles>=23.2.1
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-ulid>=2.2.0
google-api-python-client>=2.108.0
google-auth>=2.25.0
httpx>=0.26.0
